+++ pkg/compactor/lifecycle.go
@@ -1,0 +1,11 @@
+package compactor
+
+import (
+	"context"
+
+	"github.com/cortexproject/cortex/pkg/ring"
+)
+
+func (c *Compactor) OnRingInstanceHeartbeat(_ *ring.Lifecycler, ctx context.Context) {
+	c.ring.AutoForgetUnhealthy(ctx)
+}